# release-21.2: internal/batchskl: return error on index overflow