@nicktrav 2022-02-02
Follow up. There wasn't actually any use of the API that changed, so this should be a clean backport.

TFTR!